(refactor) Multiple fixes and refactors for Injective Perpetual
- Some tests still missing, but this commit is meant to unblock QA for testing

+++ hummingbot/core/gateway/gateway_http_client.py
@@ -1395,7 +1395,7 @@ class GatewayHttpClient:
         order_type: OrderType,
         price: Decimal,
         size: Decimal,
-        leverage: Decimal
+        leverage: int,
     ) -> Dict[str, Any]:
         request_payload = {
             "chain": chain,
